feat: add books total

+++ dto/base_response.go
@@ -5,5 +5,5 @@ type BaseResponse struct {
 	MessageTitle string      `json:"message_title"`
 	Message      string      `json:"message"`
 	Data         interface{} `json:"data,omitempty"`
-	Total        uint        `json:"total,omitempty"`
+	Total        int64       `json:"total,omitempty"`
 }
